- Changed sizeof(int) to sizeof(uint8_t) in all functions files

+++ src/asm/functions/and_asm.c
@@ -11,6 +11,6 @@ bool and_asm(__attribute__((unused)) int fd)
 {
 	uint8_t i = 6;
 
-	write(fd, &i, sizeof(int));
+	write(fd, &i, sizeof(uint8_t));
 	return true;
 }
