+++ src/asm/functions/ldi_asm.c
@@ -11,6 +11,6 @@ bool ldi_asm(__attribute__((unused)) int fd)
 {
 	uint8_t i = 10;
 
-	write(fd, &i, sizeof(int));
+	write(fd, &i, sizeof(uint8_t));
 	return true;
 }
