+++ src/asm/functions/xor_asm.c
@@ -11,6 +11,6 @@ bool xor_asm(__attribute__((unused)) int fd)
 {
 	uint8_t i = 8;
 
-	write(fd, &i, sizeof(int));
+	write(fd, &i, sizeof(uint8_t));
 	return true;
 }
